+++ lib/defaults.d.ts
@@ -1,5 +1,5 @@
 export declare const defaultOptions: {
     updateOnResize: boolean;
-    defaultVW: undefined;
-    defaultVH: undefined;
+    defaultVw: undefined;
+    defaultVh: undefined;
 };