+++ lib/interfaces/options.d.ts
@@ -1,5 +1,5 @@
 export interface Options {
     readonly updateOnResize?: boolean;
-    readonly defaultVW?: number;
-    readonly defaultVH?: number;
+    readonly defaultVw?: number;
+    readonly defaultVh?: number;
 }