+++ lib/native/interfaces/options.d.ts
@@ -1,0 +1,4 @@
+export interface Options {
+    readonly updateOnChange?: boolean;
+    readonly source: 'window' | 'screen';
+}