+++ lib/native/useNativeViewport.d.ts
@@ -1,4 +1,4 @@
-import { Options } from '../interfaces/options';
+import { Options } from './interfaces/options';
 export declare const useNativeViewport: (options?: Options) => {
     vw: number | undefined;
     vh: number | undefined;