+++ lib/useViewport.d.ts
@@ -2,4 +2,6 @@ import { Options } from './interfaces/options';
 export declare const useViewport: (options?: Options) => {
     vw: number | undefined;
     vh: number | undefined;
+    initialVw: number | undefined;
+    initialVh: number | undefined;
 };