+++ src/useViewport.ts
@@ -5,19 +5,28 @@ import { defaultOptions } from './defaults';
 import { Options } from './interfaces/options';
 
 export const useViewport = (options: Options = defaultOptions) => {
-  const { updateOnResize, defaultVW, defaultVH } = options;
+  const { updateOnResize, defaultVw, defaultVh } = options;
 
-  const [vw, setVW] = useState(defaultVW);
-  const [vh, setVH] = useState(defaultVH);
+  const [vw, setVw] = useState(defaultVw);
+  const [vh, setVh] = useState(defaultVh);
+
+  const [initialVw, setInitialVw] = useState(defaultVw);
+  const [initialVh, setInitialVh] = useState(defaultVh);
+
+  useEffect(() => {
+    setInitialVw(window.innerWidth);
+
+    setInitialVh(window.innerHeight);
+  }, []);
 
   useEffect(() => {
     const setSizes = () => {
       if (window.innerWidth !== vw) {
-        setVW(window.innerWidth);
+        setVw(window.innerWidth);
       }
 
       if (window.innerHeight !== vh) {
-        setVH(window.innerHeight);
+        setVh(window.innerHeight);
       }
     };
 
@@ -30,5 +39,5 @@ export const useViewport = (options: Options = defaultOptions) => {
     }
   }, [updateOnResize, vh, vw]);
 
-  return { vw, vh };
+  return { vw, vh, initialVw, initialVh };
 };